move route handler to separate 'old' file
